Move graphl fetch into library
